Implemented the quicksort used by delaunator
The results I was getting from using sort-by don't match because though the distances are sorted, the ids if the distances match are not. This makes the sort unpredictable and I haven't yet worked out a better test case than checking the results from the libraries match, perhaps I can check if the sort values are different and discard the test result?